backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
line-bot-sdk==3.5.0
stripe==7.8.0
apscheduler==3.10.4
gunicorn==21.2.0
